+++ Plugins/Plugins/ShellIntegration/ShellIntegrationPlugin.cs
@@ -16,7 +16,7 @@ namespace LenovoLegionToolkit.Plugins.ShellIntegration;
 [Plugin(
     id: "shell-integration",
     name: "Shell Integration",
-    version: "1.0.3",
+    version: "1.0.5",
     description: "Integrate Lenovo Legion Toolkit with Windows shell context menu",
     author: "LenovoLegionToolkit Team",
     MinimumHostVersion = "3.6.1",